Extract scenarios test helpers.
