+++ build-tests/api-documenter-scenarios/config/heft.json
@@ -5,13 +5,6 @@
 
   "phasesByName": {
     "build": {
-      "cleanFiles": [
-        {
-          "sourcePath": "etc",
-          "includeGlobs": ["**/*"]
-        }
-      ],
-
       "tasksByName": {
         "run-scenarios": {
           "taskDependencies": ["typescript"],
