+++ build-tests/api-documenter-scenarios/package.json
@@ -13,6 +13,7 @@
     "@microsoft/api-extractor": "workspace:*",
     "@rushstack/heft": "workspace:*",
     "@rushstack/node-core-library": "workspace:*",
-    "local-node-rig": "workspace:*"
+    "local-node-rig": "workspace:*",
+    "run-scenarios-helpers": "workspace:*"
   }
 }
