+++ build-tests/run-scenarios-helpers/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "run-scenarios-helpers",
+  "description": "Helpers for the *-scenarios test projects.",
+  "version": "1.0.0",
+  "private": true,
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "scripts": {
+    "build": "heft build --clean",
+    "_phase:build": "heft build --clean"
+  },
+  "devDependencies": {
+    "@rushstack/heft": "workspace:*",
+    "local-node-rig": "workspace:*"
+  },
+  "dependencies": {
+    "@microsoft/api-extractor": "workspace:*",
+    "@rushstack/node-core-library": "workspace:*"
+  }
+}
