+++ build-tests/run-scenarios-helpers/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/local-node-rig/profiles/default/tsconfig-base.json"
+}
