+++ common/config/rush/nonbrowser-approved-packages.json
@@ -99,11 +99,11 @@
       "allowedCategories": [ "libraries" ]
     },
     {
-      "name": "@pnpm/logger",
+      "name": "@pnpm/lockfile.types",
       "allowedCategories": [ "libraries" ]
     },
     {
-      "name": "@pnpm/lockfile.types",
+      "name": "@pnpm/logger",
       "allowedCategories": [ "libraries" ]
     },
     {
@@ -801,6 +801,10 @@
     {
       "name": "resolve",
       "allowedCategories": [ "libraries" ]
+    },
+    {
+      "name": "run-scenarios-helpers",
+      "allowedCategories": [ "tests" ]
     },
     {
       "name": "sass",
