+++ rush.json
@@ -647,6 +647,12 @@
       "reviewCategory": "tests",
       "shouldPublish": false
     },
+    {
+      "packageName": "run-scenarios-helpers",
+      "projectFolder": "build-tests/run-scenarios-helpers",
+      "reviewCategory": "tests",
+      "shouldPublish": false
+    },
 
     // "build-tests-samples" folder (alphabetical order)
     {
